OneDrive/Desktop/full stack anaytics dashboard/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic-settings==2.1.0
email-validator>=2.1.0
